refactor names of classes
